requirements.txt
fastapi==0.84.0
orjson==3.8.0
pydantic==1.10.2
uvicorn==0.18.3
uvloop==0.17.0; sys_platform == "linux" or sys_platform == "darwin"
SQLAlchemy==1.4.42
SQLAlchemy-Utils==0.38.3
asyncpg==0.27.0
psycopg2-binary==2.9.6
databases[postgresql]==0.7.0
alembic==1.8.1
pytest==7.2.0
requests==2.28.1
python-dotenv==0.21.0
httpx==0.23.1
pytest_asyncio==0.20.2
shortuuid~=1.0.11
starlette~=0.19.1